requirements.txt
boto3==1.12.43
botocore==1.15.43
docutils==0.15.2
jmespath==0.9.5
python-dateutil==2.8.1
s3transfer==0.3.3
six==1.14.0
textract-trp==0.1.3
urllib3==1.25.9
